+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## January 1st 2022
+- Setup database integration.
+- Created a generic routes files for route handling.
+- Root endpoint now returns all available endpoints.
+- Enabled file logging.
+- Update the install.db script.
+- Cleaned up the depends.
+
 ## December 30th 2021
 - Setup TypeScript.
 - Setup Express and the API backbone.